Add fetch tool to Week 01 agent
